Modulo6/proyecto_vehiculos_django/requirements.txt
django-bootstrap-v5==1.0.11
django-crispy-forms==1.14.0
crispy-bootstrap5==0.6